data-processing/requirements.txt
# Python dependencies
scikit-learn>=1.3.0  # For text similarity and clustering
numpy>=1.24.0
pandas>=2.0.0

# NLP and text processing
nltk>=3.8.1  # Natural Language Toolkit
fuzzywuzzy>=0.18.0  # Fuzzy string matching
python-Levenshtein>=0.21.0  # Fast string matching

# Geocoding
geopy>=2.4.1  # Geocoding library
googlemaps>=4.10.0  # Google Maps geocoding

# Kafka (if using Kafka from Person 1)
kafka-python>=2.0.2

# Firebase
firebase-admin>=6.2.0

# Configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0

# Utilities
python-dateutil>=2.8.2
tenacity>=8.2.3  # Retry logic
requests>=2.31.0

# Logging and monitoring
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Performance
joblib>=1.3.0  # Parallel processing
